Extract sound synthesis to separate file
