+++ src/main_note_duration.py
@@ -12,6 +12,8 @@ from mingus.core import value
 
 from midi2audio import FluidSynth
 
+import music_synthesis
+
 
 def parse_args():
     parser = argparse.ArgumentParser(
@@ -51,44 +53,4 @@ if __name__ == "__main__":
 
     print("\n".join([f"{sound.duration:.3f}: {sound.symbol}\t{round(sound.duration*8/first)/8}" for sound in sounds]))
 
-    def add_sound(t, sound, duration):
-        v = 32
-        for i in range(duration-1):
-            v = value.add(v, 32)
-        if sound.note is None:
-            print(t.add_notes(None, v))
-        else:
-            print(t.add_notes(s.symbol, v))
-
-    v_sum = sum([round(sound.duration*8/first) for sound in sounds])
-
-    t = Track()
-    t.add_bar(Bar(meter=(round(v_sum/32)+1, 1)))
-    v_bar = 32
-    for s in sounds:
-        v_act = round(s.duration*8/first)
-        
-        # while v_act > v_bar:
-        #     print(f"{v_bar}:{v_act}")
-        #     add_sound(t, s, v_bar)
-        #     v_act -= v_bar
-        #     v_bar = 32
-
-        # if v_act == 0:
-        #     continue
-
-        # print(f"{v_bar}:{v_act}")
-        add_sound(t, s, v_act)
-        # v_bar -= v_act
-
-        # if v_bar == 0:
-        #     v_bar = 32
-
-
-            
-        
-    # print(t)
-    midi_file_out.write_Track("test.mid", t)
-    fs = FluidSynth()
-    fs.midi_to_audio('test.mid', 'test.wav')
-
+    music_synthesis.wav_from_sounds(sounds, first)
